Fixed SkillCheckerMiniGame
I also added the assets for the minigame. The success and fail zone finally work properly. I input a delay for each round and a counter in the top left corner. There is a randomizer of where the success zone spawns each round. Setup a lose and win panel. I need to add the bone functionality to the script