Finally fixed the memorygame
It allows the skull heads to be detected with bone detector script. I added the skull head ui. Some features it has are scaling and turning red and also shaking for additional pleasing visuals. The ray cast of the ui images were blocking the onclick events for the minigame and is now fixed. Edited the memory game script accordingly too:)